chore: rename eip4337 files to erc4337
