+++ crates/provider/Cargo.toml
@@ -98,7 +98,7 @@ anvil-node = [
     "dep:alloy-signer-local",
 ]
 debug-api = ["dep:alloy-rpc-types-trace"]
-eip4337-api = []
+erc4337-api = []
 engine-api = ["dep:alloy-rpc-types-engine"]
 net-api = []
 trace-api = ["dep:alloy-rpc-types-trace"]
