+++ crates/provider/src/ext/mod.rs
@@ -40,7 +40,7 @@ mod txpool;
 #[cfg(feature = "txpool-api")]
 pub use txpool::TxPoolApi;
 
-#[cfg(feature = "eip4337-api")]
-mod eip4337;
-#[cfg(feature = "eip4337-api")]
-pub use eip4337::Eip4337Api;
+#[cfg(feature = "erc4337-api")]
+mod erc4337;
+#[cfg(feature = "erc4337-api")]
+pub use erc4337::Erc4337Api;
